更新 Hierarchy 类属性类型和包版本
将 Hierarchy 类中的 gc_id 和 gc_parent_id 属性的数据类型从 Int32 更改为 Int64。
在 Pek.DsMallUI.csproj 文件中，更新了 Pek.AspNetCore 和 Pek.NCubeUI 包的版本。

+++ Pek.DsMallUI/Models/Hierarchy.cs
@@ -13,7 +13,7 @@ public class Hierarchy
     /// <summary>
     /// 编号
     /// </summary>
-    public Int32 gc_id { get; set; }
+    public Int64 gc_id { get; set; }
 
     /// <summary>
     /// 名称
@@ -23,7 +23,7 @@ public class Hierarchy
     /// <summary>
     /// 父编号
     /// </summary>
-    public Int32 gc_parent_id { get; set; }
+    public Int64 gc_parent_id { get; set; }
 
     /// <summary>
     /// 类型
